updated class hierarchy for Colorizer
